refactor: Rename file manager components
