+++ source/win-main/file-manager/FileItem.vue
