+++ source/win-main/file-manager/FileTree.vue
@@ -76,7 +76,7 @@
  */
 
 import { trans } from '@common/i18n-renderer'
-import TreeItem from './tree-item.vue'
+import TreeItem from './TreeItem.vue'
 import matchQuery from './util/match-query'
 import matchTree from './util/match-tree'
 import { defineComponent } from 'vue'
